requirements.txt
cassandra-driver
click
prometheus_client
